feat: add one module to filters
