+++ src/components/ModalWindow/ModalWindow.css
@@ -3,7 +3,7 @@
 }
 
 .modal__window {
-    width: 100vh;
+    width: 100%;
     height: 100vh;
     background-color: rgba(207, 78, 78, 0.4);
     position: fixed;
